chore(playwright.yml): update package manager commands to use npm instead of pnpm chore(.gitignore): add node_modules to the list of ignored files chore(.prettierignore): add /docs and /meetings directories to the list of ignored files
The package manager commands in the playwright.yml workflow file have been updated to use npm instead of pnpm. This change ensures consistency with the package manager used in the project.

The node_modules directory has been added to the .gitignore file to exclude it from version control. This directory contains the dependencies installed by npm and should not be committed to the repository.

The /docs and /meetings directories have been added to the .prettierignore file to exclude them from formatting by Prettier. These directories contain Chingu documentation files and should not be modified by the code formatting tool.