feat(searchbox.jsx): add search functionality to recipe list
The changes include adding a search functionality to the recipe list. This allows users to search
for recipes based on a search term. The search term is passed to the FetchRecipes function which
fetches the recipes from the API based on the search term. The lodash library was added to the
project to use the debounce function which delays the processing of the search function until the
user has stopped typing for a certain amount of time. This reduces the number of API calls made and
improves performance. The SearchBox component was created to handle the search input and button. The
search term is passed up to the parent component, RecipesPage, which then passes it down to the
RecipeList component. The RecipeList component then uses the search term to fetch the relevant
recipes. The StarRating component was also updated to remove a console.log statement.

BREAKING CHANGE: when the mirage server is turned off, the data return is undefined bc of too many
requests to tasty

CH-12