refactor(layout): updates the layout wrapper
The changes were made to improve the navigation of the application by adding routing for new pages
and implementing a toggle functionality for mobile view in the Header component. New SVG and PNG
assets were added to enhance the UI. A Footer component was created to provide additional
information and a link to the GitHub repository. The App component was wrapped with BrowserRouter in
main.jsx to enable routing. A new RecipesDetail page was created to display detailed information
about recipes. The BrowserRouter in routes/index.jsx was replaced with Outlet and the Footer
component was added to display it on all pages.