chore(vite.config.js): load environment variables using dotenv in vite.config.js
The dotenv package is used in vite.config.js to these environment variables. In package.json, `npm
run cz` is removed from pre-commit hooks to streamline the commit process. In api/recipes.js, the
hardcoded API key and host are replaced with environment variables to enhance security and
flexibility. In recipe-card.jsx, the recipe card is wrapped in a Link component to prepare for
future implementation of routing to a detailed recipe page.

CH-12