style(global.css): added gradients to required components
The CSS classes in heroSection.jsx were reordered to improve readability and maintainability of the
code. The 'posts' link was commented out in constants/index.js as it is not currently in use. The
background color of the SearchBox was changed from a solid color to a gradient to enhance the visual
appeal of the component. A new CSS class, watermelon-text-gradient, was added in global.css and
applied to the text in Recipes.jsx to make the user interface more engaging and visually appealing.

CH-59