feat(assets): add new FeatureRecipe image and export it in index.js
A new image file named 'Recipe Photo.png' has been added to the assets directory. This image is
intended to be used as a featured recipe image on the website. The image has been imported and
exported in the assets index.js file to make it available for use throughout the application.