feat(api): add new tips.js file to fetch tips from Tasty API
A new file, tips.js, has been added to the API directory. This file contains a function to fetch
tips from the Tasty API. This function is asynchronous and uses axios to make a GET request to the
Tasty API. The function takes in three parameters: id, from, and size, with default values for from
and size. The function returns the data received from the API or logs an error if the request fails.
This addition will allow the application to fetch and display tips from the Tasty API, enhancing the
user experience.