refactor(recipedetails.jsx): simplify recipe retrieval logic and improve variable naming
The recipe retrieval logic has been simplified by removing unnecessary variables and conditions. The
'recipeId' variable is now a constant, and the 'FetchRecipeById' function has been renamed to
'FetchRecipeDetailsById' for better clarity. The recipe data is now directly passed to the
'RecipeDetails' component, eliminating the need for an intermediate 'recipe' variable. This makes
the code cleaner and easier to understand. The code also checks  if the  recipe  is in the cache .
If not, then it will fetch the data using the `FetchRecipeDetailsById()` API Call .

CH-91, CH-92