feat(recipedetails.jsx): add check for cachedRecipesList length before attempting to find recipe
The find method is used instead of filter to retrieve the recipe from session storage because find
returns the first element that satisfies the provided testing function, which is more efficient in
this case. A check for the length of cachedRecipesList is added to prevent unnecessary operations
when the list is empty.