feat(search.jsx): add FeatureOfTheDay component and colorful balls to the search page
The FeatureOfTheDay component is now displayed on the search page when no search term is entered.
This provides users with a default view of the feature of the day when they haven't started
searching yet. Additionally, colorful balls have been added to the page for a more engaging and
visually appealing user interface.