feat(feature-of-day.jsx): replace FeatureRecipe with FeatureChickenRecipe, add navigation to recipe
A new image, FeaturedChickenSweetPotatos.jpg, has been added to the assets. This image is now
imported and exported in the assets/index.js file. In the feature-of-the-day.jsx component, the
previous FeatureRecipe image has been replaced with the new FeatureChickenRecipe image.
Additionally, a navigation function has been added to the "See the Recipe" button, which redirects
users to the specific recipe page when clicked. This enhances the user experience by providing a
direct link to the featured recipe of the day.