feat(mirageserver): add tips.json and endpoint for tips in server.js
A new file, tips.json, has been added to the mirageServer/endpoints/tips directory. This file
contains a list of tips related to recipes. The server.js file has been updated to include a new
endpoint that serves the tips.json data. This change allows the application to provide users with
helpful tips related to the recipes, enhancing the user experience.