fix(fetchrecipes): wrap localStorage.setItem in try/catch block
When the data to be saved exceeds the browser's quota, an error is thrown. This was causing the
query to fail and no search results to be displayed. Now that the localStorage save is done inside a
try/catch block, even if the localStorage save fails, the query results are returned correctly and
the search results are displayed. This bug especially affected the iPhone and iPad, but occasionally
popped up on desktop as well.

CH-100