refactor(recipe-list.jsx): change Filters import from named to default import
The Filters component was previously imported as a named import, but it has been changed to a
default import. This change was made because the Filters component was refactored to be exported as
a default export, which simplifies the import statement and makes it more consistent with other
component imports.