feat(search.jsx): enhance page heading and adjust layout for better user experience
The page heading has been upgraded from h2 to h1 for better SEO and accessibility. The heading is
now placed above the SvgComponent for a more logical flow of information. The "Orange Ball" div's
top position in medium devices has been adjusted to improve the overall layout and visual appeal.