feat: add .env.example file with VITE_TASTY_API_KEY and VITE_TASTY_API_URL variables
The .env.example file is added to provide a template for environment variables. This helps new
developers to quickly understand the required environment variables for the project. The variables
VITE_TASTY_API_KEY and VITE_TASTY_API_URL are included in the template as they are necessary for the
application to connect to the Tasty API.